pyproject.toml
[tool.poetry]
name = "python-api"
version = "0.1.0"
description = "Building an API with Python"
authors = ["midega-g <midegageorge2@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {version="^0.115.6", extras=["all"]}
SQLAlchemy = "^2.0.37"
psycopg = {version="^3.2.4", extras=["binary"]}
bcrypt = "^4.2.1"
PyJWT = {version="^2.10.1", extras=["crypto"]}

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.0"
black = "^24.10.0"
isort = "^5.13.2"
pylint = "^3.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 80
target-version = ['py311']
skip-string-normalization = true

[tool.pylint.messages_control]
disable = [
    "missing-function-docstring",
    "missing-final-newline",
    "missing-module-docstring",
    "missing-class-docstring",
    "invalid-name",
    "no-member",
    "too-few-public-methods",
    "ungrouped-imports"
]
